// ui/src/generated/IceCandidate.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * `RTCIceCandidateInit`. An empty `candidate` = end of candidates.
 */
export type IceCandidate = { candidate: string, sdp_mid: string | null, sdp_m_line_index: number | null, username_fragment: string | null, };