// ui/src/generated/IceServer.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * `RTCIceServer`.
 */
export type IceServer = { 
/**
 * `stun:host:port`, `turn:host:port?transport=udp`, ...
 */
urls: Array<string>, username: string | null, credential: string | null, };